feat(theme): move theme file to /base foder
